requirement.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.1.7
django-crispy-forms==1.12.0
django-filter==2.4.0
future==0.18.2
html5lib==1.1
idna==3.2
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.7.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
style==1.1.0
toml==0.10.2
update==0.0.1
urllib3==1.26.6
webencodings==0.5.1
xhtml2pdf==0.2.5
